Use jest-localstorage-mock
